utils/bundle-processor/requirements.txt
jsonupdate-ng==1.0.1
jsonpath-ng==1.6.1
pyyaml==6.0.2